Improve (IMO) code structure. Abandon idea of exposing notifications as a channel - can be built atop the new wrapper easily
