+++ cmd/notification/notification.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"time"
+	"syscall"
 
 	"github.com/jazzy-crane/printer"
 )
@@ -17,55 +17,60 @@ func main() {
 		os.Exit(1)
 	}
 
-	printers := make([]*printer.Printer, len(pnames))
-	for i, pname := range pnames {
-		var err error
-		printers[i], err = printer.Open(pname)
+	multiplexed := make(chan *printer.PrinterNotifyInfo)
+
+	notifyOptions := &printer.PRINTER_NOTIFY_OPTIONS{
+		Version: 2,
+		Flags:   0,
+		Count:   1,
+		PTypes: &printer.PRINTER_NOTIFY_OPTIONS_TYPE{
+			Type:    uint16(printer.JOB_NOTIFY_TYPE),
+			Count:   uint32(len(printer.JobNotifyAll)),
+			PFields: &printer.JobNotifyAll[0],
+		},
+	}
+
+	for _, pname := range pnames {
+		p, err := printer.Open(pname)
 		if err != nil {
 			log.Println("Error printer.Open", pname, err)
 			os.Exit(1)
 		}
-	}
 
-	multiplexed := make(chan printer.PrinterNotifyInfo)
-
-	for _, p := range printers {
-		notifications, err := p.StartChangeNotifications(printer.JOB_NOTIFY_TYPE, printer.JobNotifyAll)
+		notifications, err := p.FindFirstPrinterChangeNotification(printer.PRINTER_CHANGE_ALL, 0, notifyOptions)
 		if err != nil {
-			log.Println("Error StartChangeNotifications", err)
+			log.Println("Error FindFirstPrinterChangeNotification", err)
 			os.Exit(1)
 		}
 
-		go func(p <-chan printer.PrinterNotifyInfo) {
-			for msg := range p {
-				multiplexed <- msg
+		go func(p *printer.Printer, pcnh *printer.PrinterChangeNotificationHandle) {
+			defer pcnh.Close()
+			defer p.Close()
+
+			for {
+				rtn, err := pcnh.WaitOnNotification(syscall.INFINITE)
+				if err != nil {
+					continue
+				}
+
+				if rtn != syscall.WAIT_FAILED {
+					pni, err := pcnh.FindNextPrinterChangeNotification(nil)
+					if err != nil {
+						if err != printer.ErrNoNotification {
+							log.Println("Unexpected error from FindNextPrinterChangeNotification", err)
+						}
+						continue
+					}
+					multiplexed <- pni
+				}
 			}
-		}(notifications)
+		}(p, notifications)
 	}
 
-	timeout := time.After(time.Minute)
-
-	running := true
-	for running {
-		select {
-		case <-timeout:
-			running = false
-		case pni := <-multiplexed:
-			fmt.Printf("\nNew print notification (cause 0x%X)\n", pni.Cause)
-			for _, item := range pni.Data {
-				fmt.Println(item)
-			}
-		}
-	}
-
-	for _, p := range printers {
-		if err := p.EndChangeNotifications(); err != nil {
-			log.Println("Error EndChangeNotifications", err)
-			os.Exit(1)
-		}
-		if err := p.Close(); err != nil {
-			log.Println("Error Close", err)
-			os.Exit(1)
+	for notification := range multiplexed {
+		fmt.Printf("\nNew print notification (cause 0x%X)\n", notification.Cause)
+		for _, item := range notification.Data {
+			fmt.Println(item)
 		}
 	}
 }
